Add pkg-config support
This PR implements the wishlist bug from #22. The pkg-config file format
is pretty simple, it basically provides a bunch of compiler flags to
compile and link against a library.

I changed the default INSTALL_DIR locations to relative paths, because
CMake expands relative paths to the CMAKE_INSTALL_PREFIX by default
anyway, and it makes it possible to have both the CMake config and the
pkg-config in a relocateable style.

Most of the magic happens in the foreach() loop, which is a semicolon
separated list of config values for the qhull.pc.in template in the form
NAME;DESCRIPTION